Stop pooling surface buffers
Using the fixed pool of two buffers to cache wl_buffers is pretty
pointless since buffer generally have a different size than whatever
buffers were cached.

As the pool has fixed size, it is possible to run out of buffers if
configure events arrive faster than pool buffers are marked not busy,
which can lead to protocol errors when the background surface is committed
after acknowledging a new size, but without attaching a buffer that
matches that size.